move new enum to separate file
